Changes from Code Review
